issue #21: update tests with versioning
